+++ frontend/src/lib/components/ui/dropdown-menu/dropdown-menu-trigger.svelte
@@ -4,4 +4,9 @@
 	let { ref = $bindable(null), ...restProps }: DropdownMenuPrimitive.TriggerProps = $props();
 </script>
 
-<DropdownMenuPrimitive.Trigger bind:ref data-slot="dropdown-menu-trigger" {...restProps} />
+<DropdownMenuPrimitive.Trigger
+	bind:ref
+	data-slot="dropdown-menu-trigger"
+	data-cuelume-toggle="toggle"
+	{...restProps}
+/>